Navigation 'Characters Layer' adicionado com sucesso.
