+++ app/src/main/java/com/lucascabral/marvelsuperheroes/presenter/adapter/AllCharactersAdapter.kt
@@ -1,6 +1,5 @@
 package com.lucascabral.marvelsuperheroes.presenter.adapter
 
-import android.content.Intent
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.paging.PagingDataAdapter
@@ -9,7 +8,6 @@ import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.lucascabral.marvelsuperheroes.databinding.ItemCharacterBinding
 import com.lucascabral.marvelsuperheroes.data.network.model.character.Character
-import com.lucascabral.marvelsuperheroes.presenter.view.CharacterDetailsActivity
 
 class AllCharactersAdapter(
     private val onItemClickListener: ((item: Character) -> Unit)
@@ -38,15 +36,6 @@ class AllCharactersAdapter(
                     onItemClickListener.invoke(data)
                 }
             }
-            /*itemView.setOnClickListener {
-                val directions =
-                val intent = Intent(it.context, CharacterDetailsActivity::class.java)
-                val uri = data.thumbnail.path+"."+data.thumbnail.extension
-                intent.putExtra(CharacterDetailsActivity.HERO_NAME, data.name)
-                intent.putExtra(CharacterDetailsActivity.HERO_DESCRIPTION, data.description)
-                intent.putExtra(CharacterDetailsActivity.HERO_URI, uri)
-                it.context.startActivity(intent)
-            } */
         }
     }
