+++ app/src/main/res/layout/activity_marvel_youtube.xml
@@ -9,14 +9,24 @@
     android:id="@+id/marvelYoutubeActivity"
     tools:context=".presenter.view.MarvelYoutubeActivity">
 
+    <androidx.appcompat.widget.Toolbar
+        android:id="@+id/toolbarYoutube"
+        android:layout_width="match_parent"
+        android:layout_height="?attr/actionBarSize"
+        android:background="@color/red_toolbar"
+        android:theme="@style/Theme.AppCompat.NoActionBar"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
+
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/youtubeRecyclerView"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
+        android:layout_height="0dp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/toolbarYoutube"
         tools:listitem="@layout/item_video" />
 
     <ProgressBar
