+++ app/src/main/res/layout/activity_navigation.xml
@@ -11,7 +11,7 @@
         android:id="@+id/toolbarMain"
         android:layout_width="match_parent"
         android:layout_height="?attr/actionBarSize"
-        android:background="@color/purple_500"
+        android:background="@color/red_toolbar"
         android:theme="@style/Theme.AppCompat.NoActionBar" />
 
     <fragment
